airflow/requirements.txt
jsonschema==4.23.0
pandas==1.5.3
numpy>=1.24,<1.25
SQLAlchemy>=1.4,<2.0
psycopg[binary]==3.2.9
